added rolling logs, fixed circular dependency, added exception handling and test endpoint

+++ src/main/java/com/secretcow/holygrail/entities/UserDetail.java
@@ -29,11 +29,11 @@ public class UserDetail {
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
